+++ tests/io/influxdb/test_cli.py
@@ -1,30 +1,32 @@
 # ruff: noqa: E402
 import pytest
+from click.testing import CliRunner
+from pueblo.testing.dataframe import DataFrameFactory
+
+from cratedb_toolkit.cli import cli
 
 pytest.importorskip("influxio", reason="Skipping InfluxDB tests because 'influxio' package is not installed")
 pytest.importorskip(
     "influxdb_client", reason="Skipping InfluxDB tests because 'influxdb-client' package is not installed"
 )
 
-from click.testing import CliRunner
 from influxio.model import InfluxDbAdapter
-from influxio.testdata import DataFrameFactory
-
-from cratedb_toolkit.cli import cli
 
 
 def test_influxdb2_load_table(caplog, cratedb, influxdb):
     """
-    CLI test: Invoke `ctk load table`.
+    CLI test: Invoke `ctk load table` for InfluxDB.
     """
     cratedb_url = f"{cratedb.get_connection_url()}/testdrive/demo"
     influxdb_url = f"{influxdb.get_connection_url()}/testdrive/demo"
 
-    # Populate source database with a few records worth of data.
-    adapter = InfluxDbAdapter.from_url(influxdb_url)
-    adapter.ensure_bucket()
+    # Create sample dataset with a few records worth of data.
     dff = DataFrameFactory(rows=42)
     df = dff.make("dateindex")
+
+    # Populate source database.
+    adapter = InfluxDbAdapter.from_url(influxdb_url)
+    adapter.ensure_bucket()
     adapter.write_df(df)
 
     # Run transfer command.