+++ tests/io/mongodb/test_extract.py
@@ -1,4 +1,11 @@
+# ruff: noqa: E402
 import unittest
+
+import pytest
+
+pytest.importorskip("bson", reason="Skipping tests because bson is not installed")
+pytest.importorskip("pymongo", reason="Skipping tests because pymongo is not installed")
+pytest.importorskip("rich", reason="Skipping tests because rich is not installed")
 
 import bson